Add newline at end of llmsummary
